pipeline/requirements.txt
paho-mqtt==2.1.0
influxdb-client==1.46.0
requests==2.32.3
